requirements.txt
aiohttp==3.9.1
SQLAlchemy==2.0.23
asyncpg==0.29.0
PyJWT==2.8.0
cryptography==41.0.7
python-dotenv==1.0.0
werkzeug==3.0.1